Updated Task Page
These are changes which were lost in the previous merge
All code is of full credit to Mohan Kumar and Kevin Lau, I'm just manually adding it in
This changes the appearance of the Tasks component and adds in dummy data which shows the functionality of the drag and drop 
I have also added a logout button in the header
Populate tasks with app-task-card


Merge remote-tracking branch 'origin/feature/integration' into feature/integration
Merge remote-tracking branch 'origin/feature/integration' into feature/integration

The actual merge changes are in the previous commit, however git was unhappy that I squashed it before pushing so it kept forcing me to merge. I just left in some whitespace to keep git happy.